adding new isa-validation api instead of schema based regex.
This is particularly helpful when other tools require to validate a n ISA string alone. These tools can now import `riscv_config.isa_validator` and call the `get_extension_list` to check  for errors and also return a list of all extensions present in the ISA string.

Significant change here is moving the regex from the schema to the constants file which can now be easily imported without having to load the schema yaml.